Working - passing unit tests for existing functionality
