+++ prov-api/docker-requirements.txt
@@ -14,8 +14,6 @@ cachetools
 aiocache
 types-cachetools
 prov
-# prov-db-connector (can't use original due to bug - this is forked version)
-git+https://github.com/gbrrestoration/prov-db-connector#egg=prov-db-connector
 neo4j
 networkx
 python-dotenv
