+++ prov-api/routes/admin/graph_admin.py
@@ -3,7 +3,7 @@ from ProvenaInterfaces.SharedTypes import StatusResponse, Status
 from KeycloakFastAPI.Dependencies import User
 from dependencies.dependencies import admin_user_protected_role_dependency
 from config import Config, get_settings
-from helpers.neo4j_helpers import run_query
+from helpers.prov_connector import run_query
 
 router = APIRouter()
