+++ prov-api/routes/explore/explore.py
@@ -5,7 +5,7 @@ from ProvenaInterfaces.ProvenanceAPI import *
 from ProvenaInterfaces.ProvenanceModels import *
 from ProvenaInterfaces.SharedTypes import Status
 from helpers.entity_validators import unknown_validator, RequestStyle, ServiceAccountProxy
-import helpers.neo4j_helpers as neo4j_helpers
+import helpers.prov_connector as prov_connector
 from typing import Dict, Any
 from config import Config, get_settings
 
@@ -60,7 +60,7 @@ async def explore_upstream(
     await unknown_validator(id=starting_id, config=config, request_style=request_style)
 
     # make lineage query
-    json_serialisation: Dict[str, Any] = neo4j_helpers.upstream_query(
+    json_serialisation: Dict[str, Any] = prov_connector.upstream_query(
         starting_id=starting_id,
         depth=depth,
         config=config
@@ -127,7 +127,7 @@ async def explore_downstream(
     await unknown_validator(id=starting_id, config=config, request_style=request_style)
 
     # make lineage query
-    json_serialisation: Dict[str, Any] = neo4j_helpers.downstream_query(
+    json_serialisation: Dict[str, Any] = prov_connector.downstream_query(
         starting_id=starting_id,
         depth=depth,
         config=config
@@ -170,7 +170,7 @@ async def contributing_datasets(
     await unknown_validator(id=starting_id, config=config, request_style=request_style)
 
     # make lineage query
-    json_serialisation: Dict[str, Any] = neo4j_helpers.special_contributing_dataset_query(
+    json_serialisation: Dict[str, Any] = prov_connector.special_contributing_dataset_query(
         starting_id=starting_id,
         depth=depth,
         config=config
@@ -213,7 +213,7 @@ async def effected_datasets(
     await unknown_validator(id=starting_id, config=config, request_style=request_style)
 
     # make lineage query
-    json_serialisation: Dict[str, Any] = neo4j_helpers.special_effected_dataset_query(
+    json_serialisation: Dict[str, Any] = prov_connector.special_effected_dataset_query(
         starting_id=starting_id,
         depth=depth,
         config=config
@@ -256,7 +256,7 @@ async def contributing_agents(
     await unknown_validator(id=starting_id, config=config, request_style=request_style)
 
     # make lineage query
-    json_serialisation: Dict[str, Any] = neo4j_helpers.special_contributing_agent_query(
+    json_serialisation: Dict[str, Any] = prov_connector.special_contributing_agent_query(
         starting_id=starting_id,
         depth=depth,
         config=config
@@ -299,7 +299,7 @@ async def effected_agents(
     await unknown_validator(id=starting_id, config=config, request_style=request_style)
 
     # make lineage query
-    json_serialisation: Dict[str, Any] = neo4j_helpers.special_effected_agent_query(
+    json_serialisation: Dict[str, Any] = prov_connector.special_effected_agent_query(
         starting_id=starting_id,
         depth=depth,
         config=config
